+++ apps/api/src/app/portfolio/portfolio.controller.ts
@@ -4,9 +4,12 @@ import {
   hasNotDefinedValuesInObject,
   nullifyValuesInObject
 } from '@ghostfolio/api/helper/object.helper';
+import { TransformDataSourceInRequestInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-request.interceptor';
+import { TransformDataSourceInResponseInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-response.interceptor';
 import { ConfigurationService } from '@ghostfolio/api/services/configuration.service';
 import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data.service';
 import { baseCurrency } from '@ghostfolio/common/config';
+import { parseDate } from '@ghostfolio/common/helper';
 import {
   PortfolioChart,
   PortfolioDetails,
@@ -25,17 +28,16 @@ import {
   Inject,
   Param,
   Query,
-  Res,
-  UseGuards
+  UseGuards,
+  UseInterceptors
 } from '@nestjs/common';
 import { REQUEST } from '@nestjs/core';
 import { AuthGuard } from '@nestjs/passport';
-import { Response } from 'express';
 import { StatusCodes, getReasonPhrase } from 'http-status-codes';
 
 import { PortfolioPositionDetail } from './interfaces/portfolio-position-detail.interface';
 import { PortfolioPositions } from './interfaces/portfolio-positions.interface';
-import { PortfolioService } from './portfolio.service';
+import { PortfolioServiceStrategy } from './portfolio-service.strategy';
 
 @Controller('portfolio')
 export class PortfolioController {
@@ -43,7 +45,7 @@ export class PortfolioController {
     private readonly accessService: AccessService,
     private readonly configurationService: ConfigurationService,
     private readonly exchangeRateDataService: ExchangeRateDataService,
-    private readonly portfolioService: PortfolioService,
+    private readonly portfolioServiceStrategy: PortfolioServiceStrategy,
     @Inject(REQUEST) private readonly request: RequestWithUser,
     private readonly userService: UserService
   ) {}
@@ -52,13 +54,11 @@ export class PortfolioController {
   @UseGuards(AuthGuard('jwt'))
   public async getChart(
     @Headers('impersonation-id') impersonationId: string,
-    @Query('range') range,
-    @Res() res: Response
+    @Query('range') range
   ): Promise<PortfolioChart> {
-    const historicalDataContainer = await this.portfolioService.getChart(
-      impersonationId,
-      range
-    );
+    const historicalDataContainer = await this.portfolioServiceStrategy
+      .get()
+      .getChart(impersonationId, range);
 
     let chartData = historicalDataContainer.items;
 
@@ -90,37 +90,37 @@ export class PortfolioController {
       });
     }
 
-    return <any>res.json({
+    return {
       hasError,
       chart: chartData,
       isAllTimeHigh: historicalDataContainer.isAllTimeHigh,
       isAllTimeLow: historicalDataContainer.isAllTimeLow
-    });
+    };
   }
 
   @Get('details')
   @UseGuards(AuthGuard('jwt'))
+  @UseInterceptors(TransformDataSourceInResponseInterceptor)
   public async getDetails(
     @Headers('impersonation-id') impersonationId: string,
-    @Query('range') range,
-    @Res() res: Response
-  ): Promise<PortfolioDetails> {
+    @Query('range') range
+  ): Promise<PortfolioDetails & { hasError: boolean }> {
     if (
       this.configurationService.get('ENABLE_FEATURE_SUBSCRIPTION') &&
       this.request.user.subscription.type === 'Basic'
     ) {
-      res.status(StatusCodes.FORBIDDEN);
-      return <any>res.json({ accounts: {}, holdings: {} });
+      throw new HttpException(
+        getReasonPhrase(StatusCodes.FORBIDDEN),
+        StatusCodes.FORBIDDEN
+      );
     }
 
     let hasError = false;
 
     const { accounts, holdings, hasErrors } =
-      await this.portfolioService.getDetails(
-        impersonationId,
-        this.request.user.id,
-        range
-      );
+      await this.portfolioServiceStrategy
+        .get()
+        .getDetails(impersonationId, this.request.user.id, range);
 
     if (hasErrors || hasNotDefinedValuesInObject(holdings)) {
       hasError = true;
@@ -161,26 +161,27 @@ export class PortfolioController {
       }
     }
 
-    return <any>res.json({ accounts, hasError, holdings });
+    return { accounts, hasError, holdings };
   }
 
   @Get('investments')
   @UseGuards(AuthGuard('jwt'))
   public async getInvestments(
-    @Headers('impersonation-id') impersonationId: string,
-    @Res() res: Response
+    @Headers('impersonation-id') impersonationId: string
   ): Promise<PortfolioInvestments> {
     if (
       this.configurationService.get('ENABLE_FEATURE_SUBSCRIPTION') &&
       this.request.user.subscription.type === 'Basic'
     ) {
-      res.status(StatusCodes.FORBIDDEN);
-      return <any>res.json({});
+      throw new HttpException(
+        getReasonPhrase(StatusCodes.FORBIDDEN),
+        StatusCodes.FORBIDDEN
+      );
     }
 
-    let investments = await this.portfolioService.getInvestments(
-      impersonationId
-    );
+    let investments = await this.portfolioServiceStrategy
+      .get()
+      .getInvestments(impersonationId);
 
     if (
       impersonationId ||
@@ -197,20 +198,18 @@ export class PortfolioController {
       }));
     }
 
-    return <any>res.json({ firstOrderDate: investments[0]?.date, investments });
+    return { firstOrderDate: parseDate(investments[0]?.date), investments };
   }
 
   @Get('performance')
   @UseGuards(AuthGuard('jwt'))
   public async getPerformance(
     @Headers('impersonation-id') impersonationId: string,
-    @Query('range') range,
-    @Res() res: Response
+    @Query('range') range
   ): Promise<{ hasErrors: boolean; performance: PortfolioPerformance }> {
-    const performanceInformation = await this.portfolioService.getPerformance(
-      impersonationId,
-      range
-    );
+    const performanceInformation = await this.portfolioServiceStrategy
+      .get()
+      .getPerformance(impersonationId, range);
 
     if (
       impersonationId ||
@@ -222,20 +221,19 @@ export class PortfolioController {
       );
     }
 
-    return <any>res.json(performanceInformation);
+    return performanceInformation;
   }
 
   @Get('positions')
   @UseGuards(AuthGuard('jwt'))
+  @UseInterceptors(TransformDataSourceInResponseInterceptor)
   public async getPositions(
     @Headers('impersonation-id') impersonationId: string,
-    @Query('range') range,
-    @Res() res: Response
+    @Query('range') range
   ): Promise<PortfolioPositions> {
-    const result = await this.portfolioService.getPositions(
-      impersonationId,
-      range
-    );
+    const result = await this.portfolioServiceStrategy
+      .get()
+      .getPositions(impersonationId, range);
 
     if (
       impersonationId ||
@@ -251,13 +249,12 @@ export class PortfolioController {
       });
     }
 
-    return <any>res.json(result);
+    return result;
   }
 
   @Get('public/:accessId')
   public async getPublic(
-    @Param('accessId') accessId,
-    @Res() res: Response
+    @Param('accessId') accessId
   ): Promise<PortfolioPublicDetails> {
     const access = await this.accessService.access({ id: accessId });
     const user = await this.userService.user({
@@ -265,8 +262,10 @@ export class PortfolioController {
     });
 
     if (!access) {
-      res.status(StatusCodes.NOT_FOUND);
-      return <any>res.json({ accounts: {}, holdings: {} });
+      throw new HttpException(
+        getReasonPhrase(StatusCodes.NOT_FOUND),
+        StatusCodes.NOT_FOUND
+      );
     }
 
     let hasDetails = true;
@@ -274,10 +273,9 @@ export class PortfolioController {
       hasDetails = user.subscription.type === 'Premium';
     }
 
-    const { holdings } = await this.portfolioService.getDetails(
-      access.userId,
-      access.userId
-    );
+    const { holdings } = await this.portfolioServiceStrategy
+      .get()
+      .getDetails(access.userId, access.userId);
 
     const portfolioPublicDetails: PortfolioPublicDetails = {
       hasDetails,
@@ -310,7 +308,7 @@ export class PortfolioController {
       }
     }
 
-    return <any>res.json(portfolioPublicDetails);
+    return portfolioPublicDetails;
   }
 
   @Get('summary')
@@ -318,7 +316,9 @@ export class PortfolioController {
   public async getSummary(
     @Headers('impersonation-id') impersonationId
   ): Promise<PortfolioSummary> {
-    let summary = await this.portfolioService.getSummary(impersonationId);
+    let summary = await this.portfolioServiceStrategy
+      .get()
+      .getSummary(impersonationId);
 
     if (
       impersonationId ||
@@ -341,16 +341,17 @@ export class PortfolioController {
     return summary;
   }
 
-  @Get('position/:symbol')
+  @Get('position/:dataSource/:symbol')
+  @UseInterceptors(TransformDataSourceInRequestInterceptor)
   @UseGuards(AuthGuard('jwt'))
   public async getPosition(
     @Headers('impersonation-id') impersonationId: string,
+    @Param('dataSource') dataSource,
     @Param('symbol') symbol
   ): Promise<PortfolioPositionDetail> {
-    let position = await this.portfolioService.getPosition(
-      impersonationId,
-      symbol
-    );
+    let position = await this.portfolioServiceStrategy
+      .get()
+      .getPosition(dataSource, impersonationId, symbol);
 
     if (position) {
       if (
@@ -379,19 +380,18 @@ export class PortfolioController {
   @Get('report')
   @UseGuards(AuthGuard('jwt'))
   public async getReport(
-    @Headers('impersonation-id') impersonationId: string,
-    @Res() res: Response
+    @Headers('impersonation-id') impersonationId: string
   ): Promise<PortfolioReport> {
     if (
       this.configurationService.get('ENABLE_FEATURE_SUBSCRIPTION') &&
       this.request.user.subscription.type === 'Basic'
     ) {
-      res.status(StatusCodes.FORBIDDEN);
-      return <any>res.json({ rules: [] });
+      throw new HttpException(
+        getReasonPhrase(StatusCodes.FORBIDDEN),
+        StatusCodes.FORBIDDEN
+      );
     }
 
-    return <any>(
-      res.json(await this.portfolioService.getReport(impersonationId))
-    );
+    return await this.portfolioServiceStrategy.get().getReport(impersonationId);
   }
 }