+++ apps/client/src/app/components/position/position-detail-dialog/interfaces/interfaces.ts
@@ -1,6 +1,10 @@
+import { DataSource } from '@prisma/client';
+
 export interface PositionDetailDialogParams {
   baseCurrency: string;
+  dataSource: DataSource;
   deviceType: string;
+  hasImpersonationId: boolean;
   locale: string;
   symbol: string;
 }