+++ apps/client/src/app/services/import-transactions.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { CreateOrderDto } from '@ghostfolio/api/app/order/create-order.dto';
-import { DataSource, Type } from '@prisma/client';
+import { Account, DataSource, Type } from '@prisma/client';
 import { parse } from 'date-fns';
 import { isNumber } from 'lodash';
 import { parse as csvToJson } from 'papaparse';
@@ -12,7 +12,9 @@ import { catchError } from 'rxjs/operators';
   providedIn: 'root'
 })
 export class ImportTransactionsService {
+  private static ACCOUNT_KEYS = ['account', 'accountid'];
   private static CURRENCY_KEYS = ['ccy', 'currency'];
+  private static DATA_SOURCE_KEYS = ['datasource'];
   private static DATE_KEYS = ['date'];
   private static FEE_KEYS = ['commission', 'fee'];
   private static QUANTITY_KEYS = ['qty', 'quantity', 'shares', 'units'];
@@ -23,13 +25,11 @@ export class ImportTransactionsService {
   public constructor(private http: HttpClient) {}
 
   public async importCsv({
-    defaultAccountId,
     fileContent,
-    primaryDataSource
+    userAccounts
   }: {
-    defaultAccountId: string;
     fileContent: string;
-    primaryDataSource: DataSource;
+    userAccounts: Account[];
   }) {
     const content = csvToJson(fileContent, {
       dynamicTyping: true,
@@ -38,12 +38,11 @@ export class ImportTransactionsService {
     }).data;
 
     const orders: CreateOrderDto[] = [];
-
     for (const [index, item] of content.entries()) {
       orders.push({
-        accountId: defaultAccountId,
+        accountId: this.parseAccount({ item, userAccounts }),
         currency: this.parseCurrency({ content, index, item }),
-        dataSource: primaryDataSource,
+        dataSource: this.parseDataSource({ item }),
         date: this.parseDate({ content, index, item }),
         fee: this.parseFee({ content, index, item }),
         quantity: this.parseQuantity({ content, index, item }),
@@ -53,21 +52,13 @@ export class ImportTransactionsService {
       });
     }
 
-    await this.importJson({ defaultAccountId, content: orders });
+    await this.importJson({ content: orders });
   }
 
-  public importJson({
-    content,
-    defaultAccountId
-  }: {
-    content: CreateOrderDto[];
-    defaultAccountId: string;
-  }): Promise<void> {
+  public importJson({ content }: { content: CreateOrderDto[] }): Promise<void> {
     return new Promise((resolve, reject) => {
       this.postImport({
-        orders: content.map((order) => {
-          return { ...order, accountId: defaultAccountId };
-        })
+        orders: content
       })
         .pipe(
           catchError((error) => {
@@ -90,6 +81,29 @@ export class ImportTransactionsService {
     }, {});
   }
 
+  private parseAccount({
+    item,
+    userAccounts
+  }: {
+    item: any;
+    userAccounts: Account[];
+  }) {
+    item = this.lowercaseKeys(item);
+
+    for (const key of ImportTransactionsService.ACCOUNT_KEYS) {
+      if (item[key]) {
+        return userAccounts.find((account) => {
+          return (
+            account.id === item[key] ||
+            account.name.toLowerCase() === item[key].toLowerCase()
+          );
+        })?.id;
+      }
+    }
+
+    return undefined;
+  }
+
   private parseCurrency({
     content,
     index,
@@ -108,6 +122,18 @@ export class ImportTransactionsService {
     }
 
     throw { message: `orders.${index}.currency is not valid`, orders: content };
+  }
+
+  private parseDataSource({ item }: { item: any }) {
+    item = this.lowercaseKeys(item);
+
+    for (const key of ImportTransactionsService.DATA_SOURCE_KEYS) {
+      if (item[key]) {
+        return DataSource[item[key].toUpperCase()];
+      }
+    }
+
+    return undefined;
   }
 
   private parseDate({