+++ libs/common/src/lib/interfaces/info-item.interface.ts
@@ -1,16 +1,14 @@
-import { DataSource } from '@prisma/client';
-
 import { Statistics } from './statistics.interface';
 import { Subscription } from './subscription.interface';
 
 export interface InfoItem {
   currencies: string[];
   demoAuthToken: string;
+  fearAndGreedDataSource?: string;
   globalPermissions: string[];
   isReadOnlyMode?: boolean;
   lastDataGathering?: Date;
   platforms: { id: string; name: string }[];
-  primaryDataSource: DataSource;
   statistics: Statistics;
   stripePublicKey?: string;
   subscriptions: Subscription[];